chore: refactor spells
