+++ src/lib/mechanics/atonement.ts
@@ -2,6 +2,7 @@ import { StateSpellReducer, SimState } from "../types";
 import { hasAura, numBuffsActive } from "../buff";
 import { calculateDamage } from "./damage";
 import { getMasteryPerc } from "../player";
+import { applyAura } from "./aura";
 
 /**
  * Returns the simulation state with healing and absorbs from Atonement applied
@@ -27,3 +28,19 @@ export const atonement: StateSpellReducer = (state, spell, tick): SimState => {
     healing: state.healing + calculatedDamage * 0.5 * activeAtonementCount * mastery,
   };
 };
+
+const ATONEMENT_BASE_DURATION = 15000;
+export const applyAtonement: StateSpellReducer = (state, spell) =>
+  applyAura(state, {
+    name: "Atonement",
+    applied: state.time,
+    duration: (state) => {
+      const hasSolatium = hasAura(state, "Solatium");
+      const hasIndemnity = hasAura(state, "Indemnity");
+
+      const pwsBonus = spell.name === "Power Word: Shield" && hasIndemnity ? 2000 : 0;
+      const smBonus = spell.name === "Shadow Mend" && hasSolatium ? 2000 : 0;
+
+      return ATONEMENT_BASE_DURATION + pwsBonus + smBonus;
+    },
+  });
