+++ src/lib/spells/Evangelism.ts
@@ -1,0 +1,11 @@
+import { advanceTime, evangelismExtension } from "../mechanics";
+import { Spell, SpellCategory } from "../types";
+
+export const Evangelism: Spell = {
+  category: SpellCategory.Cooldown,
+  id: 246287,
+  icon: "spell_holy_divineillumination",
+  name: "Evangelism",
+  offGcd: false,
+  effect: [evangelismExtension, advanceTime],
+};
