+++ src/lib/spells/index.ts
@@ -1,0 +1,57 @@
+import { Bloodlust } from "./Bloodlust";
+import { DivineStar } from "./DivineStar";
+import { Evangelism } from "./Evangelism";
+import { Halo } from "./Halo";
+import { Innervate } from "./Innervate";
+import { LightsWrath } from "./LightsWrath";
+import { Mindbender } from "./Mindbender";
+import { MindBlast } from "./MindBlast";
+import { Mindgames } from "./Mindgames";
+import { Penance } from "./Penance";
+import { PowerInfusion } from "./PowerInfusion";
+import { PowerWordRadiance } from "./PowerWordRadiance";
+import { PowerWordShield } from "./PowerWordShield";
+import { PowerWordSolace } from "./PowerWordSolace";
+import { PurgeTheWicked } from "./PurgeTheWicked";
+import { Rapture } from "./Rapture";
+import { Schism } from "./Schism";
+import { ScrawledWordOfRecall } from "./ScrawledWordOfRecall";
+import { ShadowCovenant } from "./ShadowCovenant";
+import { Shadowfiend } from "./Shadowfiend";
+import { ShadowFlamePrism } from "./ShadowFlamePrism";
+import { Shadowmend } from "./Shadowmend";
+import { ShadowWordDeath } from "./ShadowWordDeath";
+import { ShadowWordPain } from "./ShadowWordPain";
+import { Smite } from "./Smite";
+import { SpiritShell } from "./SpiritShell";
+
+const spells = {
+  Bloodlust,
+  DivineStar,
+  Evangelism,
+  Halo,
+  Innervate,
+  LightsWrath,
+  Mindbender,
+  MindBlast,
+  Mindgames,
+  Penance,
+  PowerInfusion,
+  PowerWordRadiance,
+  PowerWordShield,
+  PowerWordSolace,
+  PurgeTheWicked,
+  Rapture,
+  Schism,
+  ScrawledWordOfRecall,
+  ShadowCovenant,
+  Shadowfiend,
+  ShadowFlamePrism,
+  Shadowmend,
+  ShadowWordDeath,
+  ShadowWordPain,
+  Smite,
+  SpiritShell,
+};
+
+export default spells;
